ai/requirements.txt
flask==3.0.0
pandas==2.1.4
google-generativeai==0.3.2
huggingface_hub==0.19.4
pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.3
scikit-learn==1.3.2
python-dotenv==1.0.0
# prometheus-flask-exporter==0.22.4  # Optional for monitoring
